[CWS] remove usage of ipc component in probes (#45326)
### What does this PR do?

This PR simplifies the hostname fetching logic to be done once in the event monitor creation, and then provided to underlying users. Instead of having each users query the hostname on it's own using the ipc component. This is useful because it guarantees that all those components have the exact same hostname, but also because it allows the usage of the remote hostname component instead of unrolling our own code to do the same thing.

This PR will be followed up by other PRs doing similar work, this PR is only focused on simplifying the operation in the probes code.

### Motivation

### Describe how you validated your changes

### Additional Notes
