+++ pkg/eventmonitor/eventmonitor.go
@@ -17,7 +17,6 @@ import (
 
 	"github.com/DataDog/datadog-go/v5/statsd"
 
-	ipc "github.com/DataDog/datadog-agent/comp/core/ipc/def"
 	"github.com/DataDog/datadog-agent/pkg/eventmonitor/config"
 	secconfig "github.com/DataDog/datadog-agent/pkg/security/config"
 	"github.com/DataDog/datadog-agent/pkg/security/probe"
@@ -212,7 +211,7 @@ func (m *EventMonitor) GetStats() map[string]interface{} {
 }
 
 // NewEventMonitor instantiates an event monitoring system-probe module
-func NewEventMonitor(config *config.Config, secconfig *secconfig.Config, ipc ipc.Component, opts Opts) (*EventMonitor, error) {
+func NewEventMonitor(config *config.Config, secconfig *secconfig.Config, hostname string, opts Opts) (*EventMonitor, error) {
 	if opts.StatsdClient == nil {
 		opts.StatsdClient = &statsd.NoOpClient{}
 	}
@@ -221,7 +220,7 @@ func NewEventMonitor(config *config.Config, secconfig *secconfig.Config, ipc ipc
 		opts.ProbeOpts.StatsdClient = opts.StatsdClient
 	}
 
-	probe, err := probe.NewProbe(secconfig, ipc, opts.ProbeOpts)
+	probe, err := probe.NewProbe(secconfig, hostname, opts.ProbeOpts)
 	if err != nil {
 		return nil, err
 	}
