+++ pkg/eventmonitor/testutil/testutil.go
@@ -14,7 +14,6 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	ipcmock "github.com/DataDog/datadog-agent/comp/core/ipc/mock"
 	"github.com/DataDog/datadog-agent/pkg/eventmonitor"
 	emconfig "github.com/DataDog/datadog-agent/pkg/eventmonitor/config"
 	secconfig "github.com/DataDog/datadog-agent/pkg/security/config"
@@ -40,10 +39,8 @@ func StartEventMonitor(tb testing.TB, callback PreStartCallback) {
 	// Needed for the socket creation to work
 	require.NoError(tb, os.MkdirAll("/opt/datadog-agent/run/", 0755))
 
-	ipcComp := ipcmock.New(tb)
-
 	opts := eventmonitor.Opts{}
-	evm, err := eventmonitor.NewEventMonitor(emconfig, secconfig, ipcComp, opts)
+	evm, err := eventmonitor.NewEventMonitor(emconfig, secconfig, "test-hostname", opts)
 	require.NoError(tb, err)
 	require.NoError(tb, evm.Init())
 	callback(tb, evm)
