+++ pkg/security/utils/hostnameutils/hostname.go
@@ -9,8 +9,14 @@ package hostnameutils
 import (
 	"context"
 	"sync"
+	"time"
+
+	"github.com/avast/retry-go/v4"
 
 	ipc "github.com/DataDog/datadog-agent/comp/core/ipc/def"
+	pkgconfigsetup "github.com/DataDog/datadog-agent/pkg/config/setup"
+	pbgo "github.com/DataDog/datadog-agent/pkg/proto/pbgo/core"
+	"github.com/DataDog/datadog-agent/pkg/util/grpc"
 	"github.com/DataDog/datadog-agent/pkg/util/hostname"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 )
@@ -60,4 +66,35 @@ func GetHostnameWithContextAndFallback(ctx context.Context, ipcComp ipc.Componen
 	}
 	log.Infof("Hostname is: %s", hostnameDetected)
 	return hostnameDetected, nil
+}
+
+// getHostnameFromAgent attempts to acquire a hostname by connecting to the
+// core agent's gRPC endpoints extending the given context.
+func getHostnameFromAgent(ctx context.Context, ipcComp ipc.Component) (string, error) {
+	var hostname string
+	err := retry.Do(func() error {
+		ctx, cancel := context.WithTimeout(ctx, 1*time.Second)
+		defer cancel()
+
+		ipcAddress, err := pkgconfigsetup.GetIPCAddress(pkgconfigsetup.Datadog())
+		if err != nil {
+			return err
+		}
+
+		client, err := grpc.GetDDAgentClient(ctx, ipcAddress, pkgconfigsetup.GetIPCPort(), ipcComp.GetTLSClientConfig())
+		if err != nil {
+			return err
+		}
+
+		reply, err := client.GetHostname(ctx, &pbgo.HostnameRequest{})
+		if err != nil {
+			return err
+		}
+
+		log.Debugf("Acquired hostname from gRPC: %s", reply.Hostname)
+
+		hostname = reply.Hostname
+		return nil
+	}, retry.LastErrorOnly(true), retry.Attempts(maxAttempts), retry.Context(ctx))
+	return hostname, err
 }
